Agregada UI de section de perfil de empresas
